pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-cogs-optimizer"
version = "0.1.0"
description = "Middleware to optimize LLM COGS by routing between local and cloud models."
readme = "README.md"
authors = [{ name = "Charan K", email = "charan@example.com" }]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "google-cloud-aiplatform",  # For Vertex AI
    "openai",                   # For generic cloud APIs if needed
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "isort",
    "mypy"
]

[tool.setuptools.packages.find]
where = ["src"]
